Fix 全凸報告の場所を移動
