+++ src/client/convex/report.ts
@@ -1,15 +1,17 @@
 import * as Discord from 'discord.js'
 import Settings from 'const-settings'
 import Option from 'type-of-option'
+import * as util from '../../util'
 import * as status from './status'
 import * as date from './date'
+import * as lapAndBoss from './lapAndBoss'
 
 /**
  * 凸報告の管理を行う
  * @param msg DiscordからのMessage
  * @return 凸報告の実行結果
  */
-export const ConvexReport = async (msg: Discord.Message): Promise<Option<string>> => {
+export const Convex = async (msg: Discord.Message): Promise<Option<string>> => {
   // botのメッセージはコマンド実行しない
   if (msg.member?.user.bot) return
 
@@ -34,4 +36,22 @@ export const ConvexReport = async (msg: Discord.Message): Promise<Option<string>
       return 'Different format'
     }
   }
+}
+
+/**
+ * 全凸終了報告を行う
+ */
+export const AllConvex = async () => {
+  const day = await date.GetDay()
+  const state = await lapAndBoss.GetCurrent()
+
+  // 進行に報告をする
+  const channel = util.GetTextChannel(Settings.CHANNEL_ID.PROGRESS)
+  channel.send(
+    `${day}日目の全凸終了報告よ！\n` +
+      `今日は\`${state.lap}\`周目の\`${state.boss}\`まで進んだわ\n` +
+      `お疲れ様！次も頑張りなさい`
+  )
+
+  console.log('Complete convex end report')
 }
